Extract kindData determination (and the rest of essay creation) into toPostEssay function, use in both sendPost and editPost
