+++ apps/tlon-mobile/src/screens/ChannelScreen.tsx
@@ -127,6 +127,7 @@ export default function ChannelScreen(props: ChannelScreenProps) {
       store.editPost({
         post,
         content,
+        channel: channelQuery.data,
       });
       setEditingPost(undefined);
     },
