+++ apps/tlon-mobile/src/screens/PostScreen.tsx
@@ -125,6 +125,7 @@ export default function PostScreen(props: PostScreenProps) {
         post: editedPost,
         content,
         parentId: post.id,
+        channel,
       });
       setEditingPost(undefined);
     },
